add mixin to prevent vaporization of unstable portal fluid in the entire End dimension instead of just the biomes
yes, we can whitelist all end biomes, but sometimes there are fun bugs that do weird things to biomes, ex. we had main end islands biome get changed to a taiga???
I think it's fine to allow entirety of the end, not just valid biomes for it. keeping the tag as the 2nd check for being valid area, in-case we want to allow some other places too.